Repackage Android classes
