+++ android/src/main/java/com/sailthru/mobile/rnsdk/RNSailthruMobilePackage.java
@@ -1,5 +1,5 @@
 
-package com.reactlibrary;
+package com.sailthru.mobile.rnsdk;
 
 import android.content.Context;
